break out accessors.ts of metamodel.ts (using an agent)
